@@ -9,7 +9,7 @@
     "@types/react-facebook-login": "^4.1.0",
     "@types/react-router-dom": "^5.1.0",
     "@types/socket.io-client": "^1.4.32",
-    "axios": "^0.19.0",
+    "axios": "^0.20.0",
     "bootstrap": "^4.3.1",
     "jquery": "^3.4.1",
     "react": "^16.0.0",
@@ -21,7 +21,9 @@
     "react-scripts": "^3.2.0",
     "react-twitter-embed": "^3.0.3",
     "socket.io-client": "^2.3.0",
-    "typescript": "^3.7.2"
+    "typescript": "^4.0.2",
+    "use-firebase-auth": "^2.1.0",
+    "validator": "^13.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
