@@ -20,8 +20,8 @@
   "author": "Jon R Jackson",
   "license": "ISC",
   "devDependencies": {
-    "nodemon": "^1.19.4",
-    "ts-node-dev": "^1.0.0-pre.44"
+    "nodemon": "^2.0.4",
+    "ts-node-dev": "^1.0.0-pre.62"
   },
   "dependencies": {
     "@types/axios": "^0.14.0",
@@ -29,22 +29,22 @@
     "@types/express-session": "^1.15.15",
     "@types/jsonwebtoken": "^8.3.5",
     "@types/mongoose": "^5.5.29",
-    "@types/node": "^12.12.5",
+    "@types/node": "^14.10.1",
     "@types/socket.io": "^2.1.4",
-    "bcrypt": "^3.0.2",
+    "bcrypt": "^5.0.0",
     "bcryptjs": "^2.4.3",
-    "concurrently": "^5.0.0",
-    "connect-mongo": "^3.1.2",
+    "concurrently": "^5.3.0",
+    "connect-mongo": "^3.2.0",
     "dotenv": "^8.2.0",
     "express": "^4.16.4",
-    "express-session": "^1.15.6",
+    "express-session": "^1.17.1",
     "if-env": "^1.0.4",
     "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.7.8",
-    "morgan": "^1.9.1",
+    "mongoose": "^5.10.5",
+    "morgan": "^1.10.0",
     "node-tweet-stream": "^2.0.3",
     "socket.io": "^2.3.0",
-    "ts-node": "^8.4.1",
-    "typescript": "^3.7.2"
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.2"
   }
 }
