@@ -1,16 +1,15 @@
-import React, { Component } from 'react';
-// import { Link } from "react-router-dom";
+import React, { useEffect, useState } from 'react';
 
 import { AxiosResponse } from 'axios';
-import ioClient from 'socket.io-client';
-
-import { Button, Card, Container, Form, ListGroup, ListGroupItem } from 'react-bootstrap';
-
-import { Row, Col } from '../../components/Grid';
-
 import { TwitterTweetEmbed } from 'react-twitter-embed';
+import ioClient from 'socket.io-client';
+import { useFirebaseAuth } from 'use-firebase-auth';
+
+import { Button, Card, Container, Form, ListGroup, ListGroupItem, Row, Col } from 'react-bootstrap';
 
 import API from '../../utils/API';
+
+import './Yahtzee.css';
 
 const styles = {
   chatMsg: {
@@ -27,124 +26,104 @@ const styles = {
   },
 };
 
-interface YProps {
-  user: any;
-}
+function Yahtzee() {
+  const [players, setPlayers] = useState([]);
+  const [socket, setSocket] = useState({} as SocketIOClient.Socket);
+  const [incomingMsgs, setIncomingMsgs] = useState([] as any);
+  const [tweets, setTweets] = useState([] as any);
+  const [outgoingMsg, setOutgoingMsg] = useState('');
+  const [endpoint, setEndpoint] = useState('http://localhost:3001/');
+  const { user } = useFirebaseAuth();
 
-interface YState {
-  players: Array<any>;
-  socket: SocketIOClient.Socket | null;
-  incomingMsgs: Array<any>;
-  tweets: Array<any>;
-  endpoint: string;
-  outgoingMsg: string;
-}
+  useEffect(() => {
+    API.checkConnection().then((res: AxiosResponse) => console.log(res.data.msg));
+    setSocket(ioClient(process.env.NODE_ENV === 'production' ? window.location.hostname : endpoint));
+  }, [endpoint]);
 
-class Yahtzee extends Component<YProps, YState> {
-  state: YState = {
-    players: [],
-    socket: null,
-    incomingMsgs: [],
-    tweets: [],
-    outgoingMsg: '',
-    endpoint: `http://localhost:3001/`,
+  useEffect(() => {
+    console.log(socket);
+    if (Object.keys(socket).length !== 0) {
+      socket.emit('data', 'IPA');
+      socket.on('tweet', (data: any) => {
+        console.log(data);
+        setTweets((tweets: any[]) => [...tweets, data.id_str]);
+      });
+      socket.on('chat', (msg: string, chatUser: any) => {
+        var messageBody = document.querySelector('#chatBody');
+        messageBody!.scrollTop = messageBody!.scrollHeight - messageBody!.clientHeight;
+        setIncomingMsgs((incomingMsgs: any[]) => [
+          ...incomingMsgs,
+          {
+            senderInfo: chatUser,
+            msgTxt: msg,
+          },
+        ]);
+      });
+    }
+  }, [socket]);
+
+  useEffect(() => {
+    const messageBody = document.querySelector('#chatBody');
+    messageBody!.scrollTop = messageBody!.scrollHeight - messageBody!.clientHeight;
+  }, [incomingMsgs]);
+
+  const submitHandler = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
+    let msg = outgoingMsg;
+    socket &&
+      user &&
+      socket.emit('chat', msg, {
+        uid: user.uid,
+        displayName: user.displayName,
+        emailVerified: user.emailVerified,
+        photoURL: user.photoURL,
+      });
+    setOutgoingMsg('');
   };
 
-  componentDidMount() {
-    this.setState(
-      {
-        socket: ioClient(process.env.NODE_ENV === 'production' ? window.location.hostname : this.state.endpoint),
-      },
-      () => {
-        console.log(this.state.socket);
-        API.checkConnection().then((res: AxiosResponse) => console.log(res.data.msg));
-        this.state.socket && this.state.socket.emit('data', 'yahtzee');
-        this.state.socket &&
-          this.state.socket.on('tweet', (data: any) => {
-            console.log(data);
-            let tweetHistory = this.state.tweets;
-            tweetHistory.unshift(data.id_str);
-            this.setState(prevState => ({
-              ...prevState,
-              tweets: tweetHistory,
-            }));
-          }) &&
-          this.state.socket.on('chat', (msg: string, chatUser: any) => {
-            console.log(this.state);
-            let newMsgArr = this.state.incomingMsgs;
-            console.log(newMsgArr);
-            var messageBody = document.querySelector('#chatBody');
-            messageBody!.scrollTop = messageBody!.scrollHeight - messageBody!.clientHeight;
-            newMsgArr.push({
-              senderInfo: chatUser,
-              msgTxt: msg,
-            });
-            this.setState(
-              {
-                incomingMsgs: newMsgArr,
-              },
-              () => {
-                const messageBody = document.querySelector('#chatBody');
-                messageBody!.scrollTop = messageBody!.scrollHeight - messageBody!.clientHeight;
-              },
-            );
-          });
-      },
-    );
-    //  this.state.socket.close();
-  }
-
-  componentWillUnmount() {
-    this.state.socket && this.state.socket.close();
-  }
-
-  handleChange = (e: React.ChangeEvent<{ name: string; value: string }>) => {
-    const { name, value } = e.target;
-    this.setState(prevState => ({
-      ...prevState,
-      [name]: value,
-    }));
-  };
-
-  submitHandler = (e: React.MouseEvent<HTMLButtonElement, MouseEvent> | React.KeyboardEvent) => {
-    let msg = this.state.outgoingMsg;
-    this.setState(
-      {
-        outgoingMsg: '',
-      },
-      () => this.state.socket && this.state.socket.emit('chat', msg, this.props.user),
-    );
-  };
-
-  pressedEnter = (e: React.KeyboardEvent) => {
+  const pressedEnter = (e: React.KeyboardEvent) => {
     if (e.key === 'Enter') {
       e.preventDefault();
-      this.submitHandler(e);
+      let msg = outgoingMsg;
+      socket && socket.emit('chat', msg, user);
+      setOutgoingMsg('');
     }
   };
 
-  render() {
-    return (
-      <Row>
-        <Col size={'9'}>
+  return (
+    <Container
+      fluid
+      style={{
+        marginTop: '-4rem',
+        minHeight: 'calc(100vh - 112px)',
+        display: 'flex',
+        flexDirection: 'column',
+        justifyContent: 'center',
+        alignItems: 'center',
+      }}
+    >
+      <img
+        src="https://images.vexels.com/media/users/3/135811/isolated/preview/f3dc1094d770aadce0dff261623fddb6-dices-3d-icon-by-vexels.png"
+        width="300"
+        height="300"
+        className="d-inline-block align-top mainlogo"
+        alt="React Bootstrap logo"
+      />
+      <Row style={{ width: '100%' }}>
+        <Col md={9}>
           <Container
             style={{
               width: '80%',
               maxWidth: '1160px',
               background: 'white',
               minHeight: '626px',
-              margin: '40px',
+              margin: '0',
               borderRadius: '5%',
+              height: '100%',
             }}
           ></Container>
         </Col>
-        <Col size={'3'}>
-          <Card style={{ margin: '40px auto 0', maxHeight: '660px', width: '90%' }}>
-            <Card.Img
-              height={100}
-              variant="top"
-              src="https://images.vexels.com/media/users/3/135811/isolated/preview/f3dc1094d770aadce0dff261623fddb6-dices-3d-icon-by-vexels.png"
-            />
+        <Col md={3}>
+          <Card style={{ margin: '40px auto 0', minHeight: '460px', width: '100%' }}>
             <Card.Body>
               <Card.Title>Stream</Card.Title>
               <Card.Text>Players Chat Area</Card.Text>
@@ -153,7 +132,7 @@ class Yahtzee extends Component<YProps, YState> {
               id="chatBody"
               className="list-group-flush"
               style={{
-                padding: 30,
+                padding: '0.2rem',
                 border: 'solid 5px blue',
                 display: 'flex',
                 flexDirection: 'column',
@@ -162,15 +141,15 @@ class Yahtzee extends Component<YProps, YState> {
                 height: '200px',
               }}
             >
-              {this.state.incomingMsgs.length > 0 &&
-                this.state.incomingMsgs.map((msg: any, i: number) => (
-                  <ListGroupItem
-                    key={i}
-                    style={this.props.user._id === msg.senderInfo._id ? styles.chatMsg : styles.chatMsg}
-                  >
-                    {msg.senderInfo.givenName} - {msg.msgTxt}
-                  </ListGroupItem>
-                ))}
+              {incomingMsgs.length > 0 &&
+                incomingMsgs.map(
+                  (msg: any, i: number) =>
+                    user && (
+                      <ListGroupItem key={i} style={user.uid === msg.senderInfo.uid ? styles.chatMsg : styles.chatMsg}>
+                        {msg.senderInfo.displayName} - {msg.msgTxt}
+                      </ListGroupItem>
+                    ),
+                )}
             </ListGroup>
             <Card.Body>
               <Form>
@@ -178,34 +157,28 @@ class Yahtzee extends Component<YProps, YState> {
                   <Form.Label>Message</Form.Label>
                   <Form.Control
                     name="outgoingMsg"
-                    value={this.state.outgoingMsg}
-                    onChange={this.handleChange}
-                    onKeyPress={this.pressedEnter}
+                    value={outgoingMsg}
+                    onChange={(e) => setOutgoingMsg(e.target.value)}
+                    onKeyPress={pressedEnter}
                   />
                 </Form.Group>
               </Form>
-              <Button variant="primary" onClick={this.submitHandler}>
+              <Button variant="primary" onClick={submitHandler}>
                 Chat
               </Button>
+              {tweets.length > 0 && (
+                <ListGroup className="list-group-flush" style={{ minHeight: '100px', overflow: 'scroll' }}>
+                  {tweets.map((tweet: any, i: number) => (
+                    <TwitterTweetEmbed key={i} tweetId={tweet} />
+                  ))}
+                </ListGroup>
+              )}
             </Card.Body>
-            <ListGroup
-              className="list-group-flush"
-              style={{ overflow: 'auto', minHeight: '100px', maxHeight: '425px' }}
-            >
-              {this.state.tweets &&
-                this.state.tweets.map((tweet: any, i: number) => {
-                  return (
-                    <ListGroupItem key={i}>
-                      <TwitterTweetEmbed tweetId={tweet} />
-                    </ListGroupItem>
-                  );
-                })}
-            </ListGroup>
           </Card>
         </Col>
       </Row>
-    );
-  }
+    </Container>
+  );
 }
 
 export default Yahtzee;
