@@ -1,4 +1,5 @@
-import React, { ReactPropTypes, FC, ChangeEvent } from "react";
+import React, { ReactPropTypes, FC, ChangeEvent } from 'react';
+import './Form.css';
 
 interface InputProps extends ReactPropTypes {
   type: string;
@@ -7,7 +8,7 @@ interface InputProps extends ReactPropTypes {
   onChange: (event: ChangeEvent<{ name?: string; value: unknown }>) => void;
 }
 
-export const Input: FC<InputProps & any> = props => (
+export const Input: FC<InputProps & any> = (props) => (
   <div className="form-group">
     <input className="form-control" {...props} />
   </div>
