@@ -1,11 +1,6 @@
 import { Response, Request, NextFunction } from 'express';
 import { Socket } from 'socket.io';
 
-// Loading evnironmental variables here
-if (process.env.NODE_ENV !== 'production') {
-  console.log('loading dev environments');
-  require('dotenv').config();
-}
 require('dotenv').config();
 
 const express = require('express');
@@ -19,6 +14,13 @@ const io = require('socket.io');
 const tw = require('./sockets/twitterSocket');
 
 const PORT = process.env.PORT || 3001;
+
+// Loading evnironmental variables here
+if (process.env.NODE_ENV !== 'production') {
+  console.log('loading dev environments');
+  require('dotenv').config();
+}
+require('dotenv').config();
 
 // Middlewares
 app.use(morgan('dev'));
@@ -47,7 +49,7 @@ if (process.env.NODE_ENV === 'production') {
 app.use(routes);
 
 // Error handler
-app.use(function(err: Error, req: Request, res: Response, next: NextFunction) {
+app.use(function (err: Error, req: Request, res: Response, next: NextFunction) {
   console.log('====== ERROR =======');
   console.error(err.stack);
   res.status(500);
@@ -61,12 +63,12 @@ const myServer = app.listen(PORT, () => {
 const mySocket = io(myServer);
 
 // Socket scripts
-mySocket.on('connection', function(socket: Socket) {
+mySocket.sockets.on('connection', function (socket: Socket) {
   console.log('User connected');
-  socket.on('data', function(data) {
+  socket.on('data', function (data: any) {
     tw.track(data);
     console.log(`Tweeting about ${data}!`);
-    tw.on('tweet', function(tweet: any) {
+    tw.on('tweet', function (tweet: any) {
       console.log(tweet.text);
       socket.emit('tweet', tweet);
       // socket.emit('tweet', tweet);
@@ -76,11 +78,12 @@ mySocket.on('connection', function(socket: Socket) {
       tw.untrack(data);
     });
   });
-  socket.on('chat', function(text: string, user: any) {
-    console.log(`${user.googleId} - ${text}`);
-    delete user.googleId;
+  socket.on('chat', function (text: string, user: any) {
+    console.log(`${user.uid} - ${text}`);
+    delete user.uid;
     socket.emit('chat', text, user);
     socket.broadcast.emit('chat', text, user);
+    socket.broadcast.emit('chat', 'Hey!', { uid: '99999999', displayName: 'Dude!' });
   });
   socket.on('disconnect', () => {
     console.log('user disconnected');
